proptypes and cleaning up
