+++ packages/mui-base/src/MenuItem/index.tsx
@@ -3,3 +3,5 @@ export { MenuItem } from './MenuItem';
 export * from './MenuItem.types';
 
 export * from './menuItemClasses';
+
+export { unwrapMenuItemContext } from './unwrapMenuItemContext';
